requirements.txt
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.17
fastapi==0.104.1
redis
apscheduler
python-jose[cryptography]
passlib
uvicorn
python-multipart